# sprint4/API/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Toolshop API",
        "description": "Toolshop REST API technical description",
        "contact": {
            "name": "Testsmith",
            "email": "info@testsmith.io"
        },
        "version": "4.0.0"
    },
    "servers": [
        {
            "url": "https://api-v4.practicesoftwaretesting.com",
            "description": "Deployed environment"
        },
        {
            "url": "http://localhost:8091",
            "description": "Local environment"
        }
    ],
    "paths": {
...
        "/messages": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Retrieve messages",
                "description": "`user` retrieves only related messages",
                "operationId": "getMessages",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "pagenumber",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "PaginatedContactMessageResponse",
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "oneOf": [
                                                    {
                                                        "$ref": "#/components/schemas/ContactResponse"
                                                    },
                                                    {
                                                        "$ref": "#/components/schemas/ContactResponseAuthenticated"
                                                    }
                                                ]
                                            }
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1,
                                            "nullable": true
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ResourceNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Contact"
                ],
                "summary": "Send new contact message",
                "description": "Send new contact message by mail",
                "operationId": "sendMessage",
                "requestBody": {
                    "description": "Contact request object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ContactRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Result of the insert",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "AddContactMessageResponse",
                                    "properties": {
                                        "success": {
                                            "description": "",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ResourceNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    }
                }
            }
        },
        "/messages/{messageId}": {
            "get": {
                "tags": [
                    "Contact"
                ],
                "summary": "Retrieve specific message",
                "description": "Retrieve specific message",
                "operationId": "getMessage",
                "parameters": [
                    {
                        "name": "messageId",
                        "in": "path",
                        "description": "The messageId parameter in path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/ContactResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ContactResponseAuthenticated"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/ItemNotFoundResponse"
                    },
                    "405": {
                        "$ref": "#/components/responses/MethodNotAllowedResponse"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ContactRequest": {
                "title": "ContactRequest",
                "required": [
                    "subject",
                    "message"
                ],
                "properties": {
                    "name": {
                        "description": "Sender name",
                        "type": "string",
                        "maxLength": 120,
                        "example": "John"
                    },
                    "email": {
                        "description": "Required when not authenticated",
                        "type": "string",
                        "maxLength": 120,
                        "example": "john@doe.example"
                    },
                    "subject": {
                        "description": "",
                        "type": "string",
                        "maxLength": 120,
                        "example": "website"
                    },
                    "message": {
                        "description": "",
                        "type": "string",
                        "maxLength": 250,
                        "example": "Something is wrong with the website."
                    }
                },
                "type": "object"
            },
...
            "ContactResponse": {
                "title": "ContactResponse",
                "properties": {
                    "name": {
                        "description": "",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "",
                        "type": "string",
                        "example": "john@doe.example"
                    },
                    "subject": {
                        "description": "",
                        "type": "string",
                        "example": "website"
                    },
                    "message": {
                        "description": "",
                        "type": "string",
                        "example": "Something is wrong with the website."
                    },
                    "status": {
                        "description": "",
                        "type": "string",
                        "example": "NEW"
                    },
                    "id": {
                        "description": "",
                        "type": "integer",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-01 08:24:56"
                    }
                },
                "type": "object"
            },
            "ContactResponseAuthenticated": {
                "title": "ContactResponseAuthenticated",
                "properties": {
                    "name": {
                        "description": "",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "description": "",
                        "type": "string",
                        "example": "john@doe.example"
                    },
                    "subject": {
                        "description": "",
                        "type": "string",
                        "example": "website"
                    },
                    "message": {
                        "description": "",
                        "type": "string",
                        "example": "Something is wrong with the website."
                    },
                    "status": {
                        "description": "",
                        "type": "string",
                        "example": "NEW"
                    },
                    "id": {
                        "description": "",
                        "type": "string",
                        "example": "1"
                    },
                    "created_at": {
                        "type": "string",
                        "example": "2022-08-01 08:24:56"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "UnauthorizedResponse": {
                "description": "Returns when user is not authenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "UnauthorizedResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ItemNotFoundResponse": {
                "description": "Returns when the resource is not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "ItemNotFoundResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Requested item not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ResourceNotFoundResponse": {
                "description": "Returns when the resource is not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "ResourceNotFoundResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Resource not found"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
            "MethodNotAllowedResponse": {
                "description": "Returns when the method is not allowed for the requested route",
                "content": {
                    "application/json": {
                        "schema": {
                            "title": "MethodNotAllowedResponse",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Method is not allowed for the requested route"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}